Refactor create edges significantly for performance.
This acheives > 2x improvement in create edges time to original ROI, and
enables create edges on full 50T in less than 15 minutes.
